# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API для проекта Imperial",
        "title": "Imperial API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@imperial.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/v1/cars": {
            "get": {
                "description": "Возвращает список автомобилей с возможностью фильтрации и пагинации",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cars"
                ],
                "summary": "Получение списка автомобилей",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Смещение для пагинации",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Лимит для пагинации",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Фильтр по названию автомобиля",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Фильтр по ID марки автомобиля",
                        "name": "mark_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Фильтр по ID категории автомобиля",
                        "name": "category_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список автомобилей",
                        "schema": {
                            "$ref": "#/definitions/car.ListCarsResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создает новый автомобиль с указанными параметрами",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cars"
                ],
                "summary": "Создание нового автомобиля",
                "parameters": [
                    {
                        "description": "Данные для создания автомобиля",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/car.CreateCarRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Автомобиль успешно создан",
                        "schema": {
                            "$ref": "#/definitions/car.CarResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/cars/{id}": {
            "get": {
                "description": "Возвращает информацию об автомобиле по указанному ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cars"
                ],
                "summary": "Получение автомобиля по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID автомобиля для получения",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация об автомобиле",
                        "schema": {
                            "$ref": "#/definitions/car.CarResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обновляет информацию об автомобиле по указанному ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cars"
                ],
                "summary": "Обновление автомобиля",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID автомобиля для обновления",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления автомобиля",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/car.UpdateCarRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Автомобиль успешно обновлен",
                        "schema": {
                            "$ref": "#/definitions/car.CarResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удаляет автомобиль по указанному ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cars"
                ],
                "summary": "Удаление автомобиля",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID автомобиля для удаления",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Автомобиль успешно удален",
                        "schema": {
                            "$ref": "#/definitions/car.MessageResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "car.CarResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/entities.CarCategory"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "mark": {
                    "$ref": "#/definitions/entities.CarMark"
                },
                "name": {
                    "type": "string"
                },
                "only_with_driver": {
                    "type": "boolean"
                },
                "price_per_day": {
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.CarTag"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "car.CreateCarRequest": {
            "type": "object",
            "required": [
                "category_id",
                "image_url",
                "mark_id",
                "name",
                "only_with_driver",
                "price_per_day",
                "tags_ids"
            ],
            "properties": {
                "category_id": {
                    "type": "integer",
                    "example": 2
                },
                "image_url": {
                    "type": "string",
                    "example": "https://example.com/car.jpg"
                },
                "mark_id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "Toyota"
                },
                "only_with_driver": {
                    "type": "boolean",
                    "example": false
                },
                "price_per_day": {
                    "type": "integer",
                    "example": 100
                },
                "tags_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "car.ListCarsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Car"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "car.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "car.UpdateCarRequest": {
            "type": "object",
            "required": [
                "category_id",
                "image_url",
                "mark_id",
                "name",
                "only_with_driver",
                "price_per_day",
                "tags_ids"
            ],
            "properties": {
                "category_id": {
                    "type": "integer",
                    "example": 2
                },
                "image_url": {
                    "type": "string",
                    "example": "https://example.com/car.jpg"
                },
                "mark_id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "Toyota"
                },
                "only_with_driver": {
                    "type": "boolean",
                    "example": false
                },
                "price_per_day": {
                    "type": "integer",
                    "example": 100
                },
                "tags_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "entities.Car": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/entities.CarCategory"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "mark": {
                    "$ref": "#/definitions/entities.CarMark"
                },
                "name": {
                    "type": "string"
                },
                "only_with_driver": {
                    "type": "boolean"
                },
                "price_per_day": {
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.CarTag"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "entities.CarCategory": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "entities.CarMark": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "entities.CarTag": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT token must be passed with `Bearer ` prefix. Example: \"Bearer eyJhbGciOiJIUzI1NiI...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}